requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
docopt==0.6.2
idna==3.3
pipreqs==0.4.11
python-decouple==3.6
requests==2.27.1
tk==0.1.0
urllib3==1.26.9
yarg==0.1.9
